chore: align webpack hotCases
