+++ packages/rspack/tests/HotTestCasesNode.test.js
@@ -1,0 +1,8 @@
+const { describeCases } = require("./HotTestCases.template");
+
+describe("HotTestCases", () => {
+	describeCases({
+		name: "async-node",
+		target: "async-node"
+	});
+});
