--- packages/rspack/tests/HotTestCasesNode.test.ts
@@ -1,8 +1,0 @@
-import { describeCases } from "./HotTestCases.template";
-
-describeCases({
-	name: "HotTestCasesNode",
-	target: "node",
-	casesPath: "hotCases",
-	hot: true
-});
