+++ packages/rspack/tests/HotTestCasesWeb.test.js
@@ -1,0 +1,8 @@
+const { describeCases } = require("./HotTestCases.template");
+
+describe("HotTestCases", () => {
+	describeCases({
+		name: "web",
+		target: "web"
+	});
+});
