--- packages/rspack/tests/HotTestCasesWeb.test.ts
@@ -1,8 +1,0 @@
-import { describeCases } from "./HotTestCases.template";
-
-describeCases({
-	name: "HotTestCasesWeb",
-	target: "web",
-	casesPath: "hotCases",
-	hot: true
-});
