+++ packages/rspack/tests/HotTestCasesWebWorker.test.js
@@ -1,0 +1,8 @@
+const { describeCases } = require("./HotTestCases.template");
+
+describe("HotTestCases", () => {
+	describeCases({
+		name: "webworker",
+		target: "webworker"
+	});
+});
