--- packages/rspack/tests/HotTestCasesWebWorker.test.ts
@@ -1,8 +1,0 @@
-import { describeCases } from "./HotTestCases.template";
-
-describeCases({
-	name: "HotTestCasesWebWorker",
-	target: "webworker",
-	casesPath: "hotCases",
-	hot: true
-});
